GS: Move GSAdapter into GSDevice11
