+++ pcsx2/GS/Renderers/DX11/GSDevice11.cpp
@@ -22,6 +22,30 @@
 #include <VersionHelpers.h>
 #include <sstream>
 
+
+GSAdapter::operator std::string() const
+{
+	char buf[sizeof "12345678:12345678:12345678:12345678"];
+	sprintf(buf, "%.4X:%.4X:%.8X:%.2X", vendor, device, subsys, rev);
+	return buf;
+}
+
+bool GSAdapter::operator==(const GSAdapter& desc_dxgi) const
+{
+	return vendor == desc_dxgi.vendor
+		&& device == desc_dxgi.device
+		&& subsys == desc_dxgi.subsys
+		&& rev == desc_dxgi.rev;
+}
+
+GSAdapter::GSAdapter(const DXGI_ADAPTER_DESC1& desc_dxgi)
+	: vendor(desc_dxgi.VendorId)
+	, device(desc_dxgi.DeviceId)
+	, subsys(desc_dxgi.SubSysId)
+	, rev(desc_dxgi.Revision)
+{
+}
+
 GSDevice11::GSDevice11()
 {
 	memset(&m_state, 0, sizeof(m_state));
